Public release of this old code of mine
Thankfully, I documented this code as I wrote it. The only new thoughts added are the bottom section of `readme.txt`, so be sure to look there if you're interested in what I have to say on the project.

Generally, the purpose of this project was for experience and learning C, a sort of medium-sized project of a coding exercise. It's definitely not perfect, and a few bits of the code are a bit embarrassing now (comments more toward the level `i = i + 1; /* Adds one to i */` as an example), but here it is anyway.